health-bot/scheduler/requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
twilio>=9.0.0
APScheduler>=3.10.4
uvloop>=0.19.0; sys_platform != 'win32'

